index.jsからgetters,mutations,actionsを分離
