+++ src/store/getters.js
@@ -1,0 +1,4 @@
+export default {
+    name: state => state.name,
+    userCoins: state => state.coins,
+}
